requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.3.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
datasets==1.17.0
dill==0.3.4
filelock==3.4.2
frozenlist==1.2.0
fsspec==2021.11.1
huggingface-hub==0.2.1
idna==3.3
joblib==1.1.0
multidict==5.2.0
multiprocess==0.70.12.2
nltk==3.6.7
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pyarrow==6.0.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
six==1.16.0
tokenizers==0.10.3
torch==1.10.1
tqdm==4.62.3
transformers==4.15.0
typing_extensions==4.0.1
urllib3==1.26.7
xxhash==2.0.2
yarl==1.7.2
